+++ ContentPatcher/docs/release-notes.md
@@ -9,9 +9,11 @@ When releasing a format change, don't forget to update the smapi.io/json schema!
 
 
 -->
-## Upcoming release
+## 1.24.8
+Released 14 January 2021 for SMAPI 3.13.0 or later.
+
 * Fixed patches added through `Action: Include` not applied immediately in some cases.
-* Fixed `Random` values in multiplayer not in sync in Stardew Valley 1.5.5+.
+* Fixed `Random` values not in sync in multiplayer in Stardew Valley 1.5.5+.
 
 ## 1.24.7
 Released 25 December 2021 for SMAPI 3.13.0 or later.